+++ inst/include/googlepolylines/utils.hpp
@@ -1,0 +1,25 @@
+#ifndef R_GOOGLEPOLYLINES_UTILS_H
+#define R_GOOGLEPOLYLINES_UTILS_H
+
+namespace googlepolylines {
+namespace utils {
+
+// template<typename Out>
+// inline void split(const std::string &s, char delim, Out result) {
+//   std::stringstream ss(s);
+//   std::string item;
+//   while (std::getline(ss, item, delim)) {
+//     *(result++) = item;
+//   }
+// }
+// 
+// inline void split(const std::string &s, char delim, std::vector< std::string >& elems ) {
+//   //global_vars::elems.clear();
+//   split(s, delim, std::back_inserter( elems ) );
+// }
+
+} // utils
+} // googlepolylines
+
+
+#endif