+++ src/RcppExports.cpp
@@ -6,15 +6,15 @@
 
 using namespace Rcpp;
 
-// rcpp_encodeSfGeometry
-Rcpp::List rcpp_encodeSfGeometry(Rcpp::List sfc, bool strip);
-RcppExport SEXP _googlePolylines_rcpp_encodeSfGeometry(SEXP sfcSEXP, SEXP stripSEXP) {
+// rcpp_encode_sfc
+Rcpp::List rcpp_encode_sfc(Rcpp::List& sfc, bool strip);
+RcppExport SEXP _googlePolylines_rcpp_encode_sfc(SEXP sfcSEXP, SEXP stripSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::List >::type sfc(sfcSEXP);
+    Rcpp::traits::input_parameter< Rcpp::List& >::type sfc(sfcSEXP);
     Rcpp::traits::input_parameter< bool >::type strip(stripSEXP);
-    rcpp_result_gen = Rcpp::wrap(rcpp_encodeSfGeometry(sfc, strip));
+    rcpp_result_gen = Rcpp::wrap(rcpp_encode_sfc(sfc, strip));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -43,19 +43,19 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_encode_polyline
-std::string rcpp_encode_polyline(std::vector<double> longitude, std::vector<double> latitude);
+std::string rcpp_encode_polyline(Rcpp::NumericVector longitude, Rcpp::NumericVector latitude);
 RcppExport SEXP _googlePolylines_rcpp_encode_polyline(SEXP longitudeSEXP, SEXP latitudeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< std::vector<double> >::type longitude(longitudeSEXP);
-    Rcpp::traits::input_parameter< std::vector<double> >::type latitude(latitudeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type longitude(longitudeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type latitude(latitudeSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_encode_polyline(longitude, latitude));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_encode_polyline_byrow
-std::vector<std::string> rcpp_encode_polyline_byrow(Rcpp::NumericVector longitude, Rcpp::NumericVector latitude);
+Rcpp::StringVector rcpp_encode_polyline_byrow(Rcpp::NumericVector longitude, Rcpp::NumericVector latitude);
 RcppExport SEXP _googlePolylines_rcpp_encode_polyline_byrow(SEXP longitudeSEXP, SEXP latitudeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -90,7 +90,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_googlePolylines_rcpp_encodeSfGeometry", (DL_FUNC) &_googlePolylines_rcpp_encodeSfGeometry, 2},
+    {"_googlePolylines_rcpp_encode_sfc", (DL_FUNC) &_googlePolylines_rcpp_encode_sfc, 2},
     {"_googlePolylines_rcpp_decode_polyline_list", (DL_FUNC) &_googlePolylines_rcpp_decode_polyline_list, 2},
     {"_googlePolylines_rcpp_decode_polyline", (DL_FUNC) &_googlePolylines_rcpp_decode_polyline, 2},
     {"_googlePolylines_rcpp_encode_polyline", (DL_FUNC) &_googlePolylines_rcpp_encode_polyline, 2},