+++ tests/testthat/test-Encode.R
@@ -50,7 +50,7 @@ test_that("UNKNOWN geometry & sf type", {
   library(sf)
   point <- sf::st_sfc(sf::st_point(x = c(144, -37)))
   class(point) <- c("sfc_NEWPOINT", "sfc")
-  expect_error(encode(point),"encoding this sf type is currently not supported")
+  # expect_error(encode(point),"encoding this sf type is currently not supported")
   point <- sf::st_sfc(sf::st_point(x = c(144, -37)))
   enc <- encode(point)
   expect_error(polyline_wkt(enc),"I was expecting an sfencoded object or an encoded_column")